+++ domain/src/main/kotlin/com/deividasstr/domain/entities/ConsumedSweet.kt
@@ -1,3 +1,3 @@
 package com.deividasstr.domain.entities
 
-data class ConsumedSweet(val id: Long = 0, val sweetId: Long, val g: Int, val date: Long)
+data class ConsumedSweet(val id: Long = 0, val sweetId: Long, val g: Long, val date: Long)
