+++ domain/src/main/kotlin/com/deividasstr/domain/entities/Sweet.kt
@@ -3,7 +3,7 @@ package com.deividasstr.domain.entities
 data class Sweet(
     val id: Long,
     val name: String,
-    val calsPer100: Double,
+    val calsPer100: Long,
     val fatG: Double,
     val carbsG: Double,
     val sugarG: Double,
