+++ ui/src/main/kotlin/com/deividasstr/ui/base/models/ConsumedSweetUi.kt
@@ -2,7 +2,7 @@ package com.deividasstr.ui.base.models
 
 import com.deividasstr.domain.entities.ConsumedSweet
 
-data class ConsumedSweetUi(val id: Long, val sweetId: Int, val g: Int, val date: Long) {
+data class ConsumedSweetUi(val id: Long, val sweetId: Int, val g: Long, val date: Long) {
 
     constructor(sweet: ConsumedSweet) : this(sweet.id, sweet.sweetId.toInt(), sweet.g, sweet.date)
 }
