+++ ui/src/main/kotlin/com/deividasstr/ui/base/models/SweetUi.kt
@@ -6,7 +6,7 @@ import com.deividasstr.ui.features.sweetdetails.SweetRating
 data class SweetUi(
     val id: Long,
     val name: String,
-    val calsPer100: Double,
+    val calsPer100: Long,
     val fatG: Double,
     val carbsG: Double,
     val sugarG: Double,
